Integrated ranking algorithm.

+++ db/migrate/20110921171047_create_items.rb
@@ -1,11 +1,11 @@
 class CreateItems < ActiveRecord::Migration
   def change
     create_table :items do |t|
-      t.integer :speed
-      t.integer :price
-      t.integer :memory
-      t.integer :hard
-      t.integer :optical
+      t.float :speed
+      t.float :price
+      t.float :memory
+      t.float :hard
+      t.float :optical
       t.string :name
       t.string :description
       t.string :product_url
